Programming Task Part 1 – Completed as Coursework
You have been contacted by the airline “Queasy Jet” who have a new
slogan: “The less

you weigh, the less you pay” and are planning to implement a possibly
controversial fare

system where the fare paid is based on a passenger’s total weight
(passenger plus

luggage). You are to design, code and test part of a simple system using
the classes

listed and detailed below which will allow the airline to evaluate this
idea. This system

will form the basis for a test under exam conditions during which you
will be asked to

modify and extend your system. You must therefore ensure that you design
the system

according to the specification below and that the code you write should
be neat, correctly

indented and comply with Java style guidelines (as taught during the
course). You must

develop and submit your system as a BlueJ project. This work itself is
not marked but

the demo (test) will be, so you need to complete it yourself and
understand the code so

you can modify it. Please write your last-name, first-name and SRN
number at the top of

each class as comment. This must be completed as an individual piece of
work –

each student works independently. Do not share your work with others.
Your

code may be submitted to a comparison system to test for similarity with
other

code.

In order to run and demonstrate your system you will need to add a
number of

passengers. This must be done in the BookingAdmin class.